requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
APScheduler==3.10.4
asyncpg==0.29.0
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
Faker==27.0.0
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
redis==5.0.8
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
yarl==1.9.4
